requirements.txt
asgiref==3.7.2
Django==4.2.6
django-model-utils==4.3.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
packaging==23.2
pdf2image==1.16.3
pdfminer==20191125
Pillow==10.0.1
pycryptodome==3.19.0
pytesseract==0.3.10
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
openai
python-decouple